+++ portlets/reports-portlet/docroot/WEB-INF/src/com/liferay/reports/service/persistence/DefinitionPersistenceImpl.java
@@ -357,7 +357,9 @@ public class DefinitionPersistenceImpl extends BasePersistenceImpl<Definition>
 						definition.getPrimaryKeyObj());
 			}
 
-			session.delete(definition);
+			if (definition != null) {
+				session.delete(definition);
+			}
 		}
 		catch (Exception e) {
 			throw processException(e);
@@ -366,7 +368,9 @@ public class DefinitionPersistenceImpl extends BasePersistenceImpl<Definition>
 			closeSession(session);
 		}
 
-		clearCache(definition);
+		if (definition != null) {
+			clearCache(definition);
+		}
 
 		return definition;
 	}
